Rename modules for consistency
