+++ deduce/redactor.py
