+++ tests/unit/test_annotation_proc.py
@@ -1,6 +1,6 @@
 import docdeid as dd
 
-from deduce.annotation_processing import (
+from deduce.annotation_processor import (
     CleanAnnotationTag,
     DeduceMergeAdjacentAnnotations,
     PersonAnnotationConverter,
